+++ services/clients/kandinsky.py
@@ -35,8 +35,7 @@ import aiohttp
 from loguru import logger
 from PIL import Image
 
-from services.clients import ITextToImageClient
-from services.protocols import IModelsRepo
+from services.protocols import IModelsRepo, ITextToImageClient
 from utils.config import KandinskyConfig
 from utils.models_repo import ModelsRepo
 from utils.retry import retry_standard